+++ toolkit/cli/commands/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod address_association_signatures;
 pub mod key_params;
 pub mod registration_signatures;
 pub mod signing;