+++ changelog.md
@@ -58,6 +58,7 @@ be set through `MC__SLOT_DURATION_MILLIS` environment variable.
 
 ## Added
 
+* command `sign-address-association` and `address-associations` pallet.
 * Command `smart-contracts reserve init`, `create`, `deposit`, `release`, `handover` and `update-settings`
 * Command `smart-contracts governance update`
 * Ogmios client backed by jsonrpsee `WsClient`